claygent/src/claygent/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pydantic==1.8.2
jinja2==3.0.1
python-dotenv==0.19.0